60 telegram knowledge/bot/requirements.txt
python-telegram-bot==13.12
requests
